Moves locale module into admin namespace
