+++ lib/alchemy/engine.rb
@@ -23,6 +23,7 @@ require 'turbolinks'
 require 'userstamp'
 
 # Require globally used Alchemy mixins
+require_relative './admin/locale'
 require_relative './auth_accessors'
 require_relative './cache_digests/template_tracker'
 require_relative './config'
@@ -34,7 +35,6 @@ require_relative './filetypes'
 require_relative './forms/builder'
 require_relative './hints'
 require_relative './i18n'
-require_relative './locale'
 require_relative './logger'
 require_relative './modules'
 require_relative './mount_point'
